Add sessionId to partitions map into the Schema
When we send a session init message we a new entry into this map
with the sessionId as the key and the list of partitions which this
instance of the session manager is assigned to. This way the gateway
can use the sessionId to route the message back to the LinkManager
with the correct in memory state (cryptographic keys).