CORE-1346 Implement message sending and receiving in the linkmanager
When a outbound message is processed the LinkManager adds a start
marker to the markers topic for that message. When this message is
recieved an Ack message is sent back to the sending linkmanager.
Once the Ack is recieved the LinkManager adds a done marker to the
markers topic. The markers AVRO schema was added as a part of
CORE-1959 along with the schema for the Ack message. This markers
topic will be monitored by the DeliveryTracker which will be implemented
as a part of CORE-1286.