CORE-1342 Implement session management in link manager (#59)
* Started implementation of session management in LinkManager

Still lots to do:
1. MessageTypes (for comms between Linkmanager and gateway are not
in Avro
2. Proper Networking support (at the moment we are using strings
as a source destination for testing this needs to use a more
realistic network interface).
3. Only implemented Authentication (not encrytption + authentication)
4. Messages are not sent to Kafka
5. In memory message replays (so if session messages are dropped
we can recover).

* Move data to Avro + more releastic integration with NetworkMap

Moved data to Avro (FlowMessage for statemachine to linkmanager
message + LinkManagerToGatewayHeader for linkmanger to gateways
comms). This needs to be wired into Kafka still. Also built a
more releastic interface with the network map (SessionNetworkMap).

* Generalize SessionManager to multiple Holding Identities.

Also minor rename of types in Avro.

* Move link-manager to components

* Fix fallout from rebase (max message size).

* Connect the SessionManagers to Subscription Factories

This is completely untested + LinkManager also needs to be started
by a LifeCycleEvent.

* Fix Detekt

* Improvements from Code review

Merged SessionManagerInitiator and SessionManagerResponder into
SessionManager. Moved lots of message handleing into Messaging
and the queues for pending sessions into LinkManager. LinkManager
now implements Lifecycle to handle LifeCycle events.

* Improve build script + add docs + rename NetMap

* Fixes to the message structure based on Code review

Changed sessionManager to extract public key from
responderHelloMessage as it is no longer duplicated in Step2Message

* Support Encryption + Authentication

This needs testing

* Respond to comments and further testing

* Test + improve SessionManger error handling if Session not pending

* Pull out SignData from NetworkMap to crypto service.

Also increase test coverage of SessionManager.

* Finished testing session manager interacting with the networkMap

* Fix detekt and tidy code. Also send messages to the correct topic.

* Rename function to be clearer.

* Improve logging

* Remove SNI from NetworkMap and add logic for calculating SNI to LinkManager

* Added tests of LinkManager components

OutboundMessageProcessor and PendingSessionMessageQueues

* Fix fallout from rebase.

* Add test of InboundMessage processor.

* Handle session message duplication correctly

Added a test of this behavour and reduced duplication in the
testcode.

* Add check of Authentication failure during session negotatiation

Added a test to check this happens.

* Fix detekt

* Test messages are dropped with logging when Authentication Fails

* Add .p2p.corda.net suffix to SNI for classic Corda Identity

This is intentionally different to Corda 4.

* Use public key hash in signing service

The public key has should be use to find the correct private key.

* Move topic names to new module called topic-schema.

This will eventually contain all the config files nessary to start
Kafka and create the topics. It will probably also contain some
code to generate EventLogSubscription config from these config
files.

* Improve how the SNI is calculated.

For Corda 5 Identities handle a full hostname.

* Respond to review comments in LinkManager

* Improve warning messages

* Update SNI calculation and fix tests

* Add type field to Holding Identity

This can either be CORDA 5 or classic Corda identity.

* Update session key to not allow nullable holdingId add extra id field.

* Fix fallout from rebase

Event log record can produce null values now.

* Route queued messages to the Gateway (the topic was wrong).

* Improve error handling in the SessionManager

* Add test that SniCalculator is future proof

* Add Lock to SessionManager to fix race

This was between session negotiation and queueing messages for
pending sessions.

* Add sessionId to partitions map into the Schema

When we send a session init message we a new entry into this map
with the sessionId as the key and the list of partitions which this
instance of the session manager is assigned to. This way the gateway
can use the sessionId to route the message back to the LinkManager
with the correct in memory state (cryptographic keys).

* Add unit test of InboundAssignmentListener + fix detekt

* Instead of adding SNI to LinkOutMessage add x500Name + IdentityType

This way the Gateway can calculate the SNI instead of the
LinkManager.

* Rename IdentityType to NetworkType

* Rework LinkManagerNetworkMap.kt to work with MemberInfo

Removed NetworkType from HoldingIdentity in both LinkMangerNetworkMap
and the message schema. Added new method getNetworkType which gets
the network type for a HoldingIdentity

* Fix race condition in tests which check logging

Also simplified some test code.