CORE-1934 Minor improvements to the LinkManager (follow up from CORE-1342).
Added a start/stop lock to the LinkManager so it can be started +
stopped in a thread safe way.
The OutboundMessageProcessor now discards the messages in the
LinkManager if the InboundMessageProcessor is not subscribed to
any partitions. As in this case we can never route a message back
to the correct session.