CORE-1959 AVRO schema for p2p out markers
Add AVRO data type for FlowMessageMarker. This will latter be
used by the LinkManager to track delivery of messages. Added
LinkManagerSentMarker which will be persisted when a message
is received by the LinkManager from the statemachine.
Added LinkManagerRecievedMarker which will be persisted when a
successful delivery acknowledgement is received by the LinkManager.