CORE-1959 AVRO schema for p2p out markers (#105)
* CORE-1959 AVRO schema for p2p out markers

Add AVRO data type for FlowMessageMarker. This will latter be
used by the LinkManager to track delivery of messages. Added
LinkManagerSentMarker which will be persisted when a message
is received by the LinkManager from the statemachine.
Added LinkManagerRecievedMarker which will be persisted when a
successful delivery acknowledgement is received by the LinkManager.

* Added extra MarkerTypes which will be used latter.

Added docs for each Marker to describe where they will be used.

* Update AVRO java doc after code review

Simplified the descriptions.

Co-authored-by: Dimos Raptis <dimos.raptis@r3.com>

* Update AVRO javadoc and regenerate AVRO from schema

* Spell received correctly in file

Co-authored-by: Dimos Raptis <dimos.raptis@r3.com>
Co-authored-by: Denis Rekalov <denis.rekalov@r3.com>