CORE-1959 Avro schema changes need for message acknowledgement (#109)
Added LinkManagerPayload in CORE-1346 this will become the payload
of the authenticated or authenticated encrypted data message.
LinkManagerPayload contains a union of FlowMessageAndKey and
MessageAck. As (at this level) the LinkManager can send either a
FlowMessage or a successful delivery acknowledgment to another
LinkManager. FlowMessageAndKey contains a FlowMessage and the original
key on p2p.out topic. This will let (in CORE-1346) the receiving
LinkManager write the FlowMessage to p2p.in with original key.

Co-authored-by: Denis Rekalov <denis.rekalov@r3.com>