CORE-1346 Implement message sending and receiving in the linkmanager (#119)
* CORE-1346 Implement message sending and receiving in the linkmanager

When a outbound message is processed the LinkManager adds a start
marker to the markers topic for that message. When this message is
recieved an Ack message is sent back to the sending linkmanager.
Once the Ack is recieved the LinkManager adds a done marker to the
markers topic. The markers AVRO schema was added as a part of
CORE-1959 along with the schema for the Ack message. This markers
topic will be monitored by the DeliveryTracker which will be implemented
as a part of CORE-1286.

* Improve type saftely in LinkManager

Convert some union Avro types to Sealed Classes (as Avro union
feilds create java fields with type Object).

* Tidy up LinkManagerTest

* Add a test for the Avro sealed classes

* Use the outbound session to decrypt/MAC an ACK message.

Previously we were using inbound session (which was wrong).
Now we can use the sessionId to work out if a message is an ACK or
FlowMessage. Removed the now unnecessary type (LinkManagerPayload)
from the AVRO schema.

* Test ACKs are dropped on inbound sessions and FlowMessages on outbound sessions

Co-authored-by: Denis Rekalov <denis.rekalov@r3.com>