CORE-2242 Add Diffie Hellman key exchange to the LinkManager (#160)
* CORE-2242 Add Diffie Hellman key exchange to the LinkManager

Updated the LinkManager to do all steps of Diffie Hellman key
exchange. Perviously the first stage was done in the Gateway (this
code will be removed as a part of CORE-2241).

* Update corda-api version + improve tests + tidy up

* Use correct keys in tests + fix docs in NetworkMap + fix gradle consts

* Change AuthenticationProtocolResponder to use key hash in Hello instead of Handshake

Updated the tests + the network map interface. This makes both
the NetworkMap interface and SessionManagerImpl simpler.

* Add extra check to AuthenticationProtocolResponder.verify

Throw exception in if key used in verify hashes differently to the
hash in InitiatorHelloMessage. This should prevent user error.

* Rename test in AuthenticationProtocolFailureTest to be less confusing

Co-authored-by: Dimos Raptis <dimos.raptis@r3.com>

* Make comment clearer in the test

Co-authored-by: Dimos Raptis <dimos.raptis@r3.com>