CORE-2617 File based `ConfigurationReadService`
Create `FileConfigurationReadServiceImpl` that provides a file based
implementation of `ConfigurationReadService`.

It takes in the path of a file it should read config from. Applications
or other components will pass it the path to the config file. The path
is passed into `bootstrapConfig` where the service will then parse the
configuration from `conf` file the path points to.