CORE-2617 File based `ConfigReader` (#248)
Create `FileConfigReaderImpl` as a library. This library implements the 
`:libs:configuration:configuration-read` interfaces and can be used
by an application by specifying it as a `runtimeOnly` dependency.

It is accessed via the 
`:components:configuration:configuration-read-service` as it simply
provides a alternative `ConfigReader` implementation that the
component uses.

The `FileConfigReaderImpl` looks at the `bootstrapConfig` to extract
the name of the file that it should parse and store configuration from.