CORE-1347 Smart Session Refresh in the LinkManager
Add a HeartbeatManager which periodically sends messages between
peers for each session. Messages are not acknowledged within a
time window then Sessions are torn down and a new session will
be negotatiated between the peers. This mechanism is used to
recover from scenarios where the peer's Link Manager crashes
and a session is lost.