Authorize on all paths using wildcard
Authorize all paths and add logic that checks the HTTP method/verb and
the path and gives them to the authorization code. These paths and verbs
are then used to determine if an incoming call is allowed or not.

Contains much less code than previous solution and is easier to
understand.