CORE-1347 Smart Session Refresh in the LinkManager (#275)
* CORE-1347 Smart Session Refresh in the LinkManager

Add a HeartbeatManager which periodically sends messages between
peers for each session. Messages are not acknowledged within a
time window then Sessions are torn down and a new session will
be negotatiated between the peers. This mechanism is used to
recover from scenarios where the peer's Link Manager crashes
and a session is lost.

* Fix bug (was no longer stripping key from message sent back to app)

Also fixed all the tests so they pass.

* Update LinkManagerTest to take into account HeartbeatManager

* Update SessionManagerTest to take into account HeartbeatManager

* Refactor HearbeatManager to schedule separately timeout + send heartbeat

Updated HeartbeatManager API to be simpler.

* Remove SessionManager from PendingSessionMessageQueuesImpl

This was unused and caused an ugly circuler dependency.

* Move HeartbeatManager into SessionManager

* Respond to Dimos' review comments

* Fix fallout from Reabse and rework the SessionManagerTest

* Minor refactor based on review comments

* Simplify SessionManager slightly

* Fix final (for now) review comments

* Use Mockito instead of implementing class

* Fix detekt and add a final few tests

* Change timeout to be in milliseconds not seconds

This makes the tests run in a more sensible time.

* Fix flakey test

* Respond to review comments on implementation

Track sessions by id and not key.

* Update tests based on review comments

* Fix Detekt: Remove unused imports