CORE-2436  POC loading cpi-specific crypto digest classes.
This is the first commit towards building the virtual node manager
and other infrastructure.  It demonstrates how we can load and
register cpi-specific custom classes with the underlying runtime
and unload them as necessary.

Use a mediator to combine crypto and sandbox context services

There are some integration tests that demonstrate this as a POC.

Naming is complete open to change (reminder:  refactoring in Intellij
is simple and mechanical).

Not wired into the rest of the virtual-node yet.  That will be addressed
in a further commit.

Unit tests implicitly unload sandbox groups `@After` a test is run.

We also explicitly unload sandbox groups to test the "unloading" of a CPI, and hence
the custom crypto being 'deregistered' from the process.

Also fixes a minor "bug" in the install-legacy code the prevents us loading the same CPI multiple times.