CORE-2842 Permission Storage Reader
The Permission Storage Reader is responsible for reading data from the
database and publishing it to the message bus.

The Permission Storage Reader consists of the
`:permission-storage-reader-service` component and
`:permission-storage-reader` + implementation libraries.

## Permission Storage Reader library

The `PermissionStorageReader` exposes a `publish` function in
preparation of integration with the `PermissionStorageWriter`, which can
be called to retrieve specific records from the database that were
updated and publish them to the message bus.

It also publishes data when the `PermissionStorageReader` is started.

This publishing logic uses the `PermissionCache` to diff what exists in
the message bus against what is in the database.

If the diff shows a record that exists in the topic but not in the
database, then `null` is sent to the topic to clear it out.

If the diff shows a record exists in the database then it is sent to the
topic, overriding the existing record or creating a new entry.

## Permission Storage Reader component

The `PermissionStorageReaderService` component handles the lifecycle and
publisher creation on behalf of the `PermissionStorageReader` library.

The `PermissionCacheService` listens to status changes from the
`PermissionCacheService` to determine when to properly start itself. The
service is only fully started after the `PermissionCacheService` is
started.