CORE-2871 Add `PermissionManagementRequest` schema (#121)
Add a top level schema that will cover all permission management
requests by using a union type of the `request` field. Other specific
permission management requests can then be added as needed in the
future.

Move `requestUserId` to this schema as it will be required for all types
of requests.