CORE-2966  Virtual node info service implementation.
Implementation of a compacted queue reader that works
with a basic `VirtualNodeInfo` Avro definition that we
expect to be further refined.

Fixed up packages to just be 'virtualnode'.

More unit tests for component to attempt to test start up ordering
and arrival of configuration.

Address issues with delayed receipt of configuration from the
messaging layer.

Add additional tests to check how the classes behave if the config
is not available.

As discussed in original PR/Teams, the virtual node "context" is the key that is passed
to get a sandbox group context of a specific type.

Get by short hash, return list rather than single object