CORE-2966  Virtual node info service implementation.
Implementation of a compacted queue reader that works with a basic
 `VirtualNodeInfo` Avro definition that we expect to be further refined.

Address issues with delayed receipt of configuration from the
messaging layer by communicating using events and using the coordinator.

Add additional tests to check how the classes behave if the config
is not available.

Only permit one short hash per holding identity (i.e. we don't handle
collisions of short hashes).

If, at the time of creation of a _new_ `HoldingIdentity` the "short hash"
collides with one that already exists, then _that part_ of the code should
reject the creation of the addition identity.

We still perform a cursory check to see that if we're putting a `VirtualNodeInfo`
into the internal map with an existing "short hash", that short hash points to
the same `HoldingIdentity` in the existing `VirtualNodeInfo`, otherwise we
raise an exception.