CORE-3096 CPI Info service
For a given CPI Identifier, the service will return the CPI Metadata.

Very similar to the VirtualNodeInfo service component.

Checks to see if the CPI.Identifier that is the 'key' in the compacted topic, matches
the CPI.Identifier within the CPI.Metadata when we add a snapshot or new value. If the
check fails, we set the component to ERROR. We do the same in the VirtualNodeInfo
service.