CORE-3026 Permission RPC Ops Service (#501)
* CORE-3026 Permission RPC Ops Service
This component listens to status updates from the permission service. When UP, it will create the endpoints for User, Group and Role operations. It gets a reference to the permission manager and permission validator and sets these on the endpoint to be used. Certain operations may require the permission validator for additional validation hence why it is present. The endpoint delegates to the permission manager to perform read and write operations. This component should be included in the HTTP Rpc gateway component and exposes a getter to obtain all permission rpcOps endpoints when the service is UP.

* CORE-3027 Add tests to setting manager/validator to null when permission service is down. Fix a few detekt errors.

* CORE-3026 Permission RPC Ops Service - make endpoint implement PluggableRpcOps, remove virtualNodId from endpoint and request dtos and types.

* CORE-3026 Permission RPC Ops Service - fix unit test

* CORE-3026 Make endpoints an OSGi component so they automatically get picked up by the PluggableRPCOps discovery logic.
Adding permissionManager and permissionValidator as vars to interface to allow us to set these by the component when permission service is ready.

* CORE-3026 Remove endpoint impl export

* CORE-3026: Illustrate proposed changes for PR #501

In particular:
1. Delete module `permission-validation` and move interfaces into module `permission-manager` and a dedicated package.
Having two separate modules is logical, but rather inconvenient as 3rd module will have to be created for `PermissionService`.
2. Delete module `:permission-rpcops-service`as  redundant
3. Create `PermissionService` interface in the `:lib` area
4. Tidy-up `UserEndpoint` interface
5. Tidy-up `UserEndpointImpl`

* CORE-3026: Make `UserEndpointImpl` a component

* CORE-3026: Restore `:permission-validation` module

As it is kind of "innocent" if we are not introducing `PermissionService`.

Co-authored-by: Viktor Kolomeyko <Viktor.Kolomeyko@r3.com>