Only permit one short hash per holding identity.
We will only allow one "short hash" representation per holding identity.

If, at the time of creation of a _new_ `HoldingIdentity` the "short hash" collides with one that already
exists, then _that part_ of the code should reject the creation of the addition identity.

We still perform a cursory check to see that if we're putting a `VirtualNodeInfo` into the internal map
with an existing "short hash", that short hash points to the same `HoldingIdentity` in the existing
`VirtualNodeInfo`, otherwise we raise an exception.