CORE-3175 Prevent LinkManager sessions from repeatidly timing out with same id
This was caused by destroyOutboundSession not removing the session
negotiation messages from the InMemorySessionReplayer.
After the HeartbeatManager times out the InMemorySessionReplayer added
the session back into the HeartbeatManager. The SessionManager
doesn't know about this and so makes another session if we get a
new message to the same peer (or the DeliveryTracker replays the
original app message). I have modified the HeartBeatManager to remove
the all session messages from the replayer.