CORE-3166  Report when an Application subclass has not been instantiated.
The `?.startup(args)` call was silently continuing if the LHS was null.

In the case where we have:

    @Component class Mine @Activate constructor(
       @Reference param1: Param1,
       @Reference param2: Param2,
       @Reference param3: Param3
    ) : Application

We may have a service _reference_ but not a service (instance) in the OSGi registry and the
application would previous silently fail to start prior to this commit.

We now log an error and recommend the removal of all parameters, and add them back one at a time
to identify the source of the error.