CORE-3175 Prevent link manager sessions from repeatedly timing out with same ID (#545)
This was caused by destroyOutboundSession not removing the session negotiation messages from the InMemorySessionReplayer. After the HeartbeatManager times out the InMemorySessionReplayer added the session back into the HeartbeatManager. The SessionManager doesn't know about this and so makes another session if we get a new message to the same peer (or the DeliveryTracker replays the original app message). I have modified the HeartBeatManager to remove the all session messages from the replayer.

I have also changed the behaviour of destroyOutboundSession to publish a tombstone record into SESSION_OUT_PARTITIONS. So the the Gateway eventually start to drop messages from destroyed sessions.

I have added the following new tests:
when messages already queued for a peer, if we applyNewConfiguration then NewSessionNeeded, when applyNewConfiguration, after the inbound session is established, the session is removed and when a data message is sent, heartbeats are sent, this stops if the session manager gets a new config.

I have updated when a data message is sent, heartbeats are sent, if these are not acknowledged the session times out to check for a tombstone record.