CORE-1546 Flow sandbox friendly dependency injection (#488)
Adds dependency injection of platform services and initialises sandbox's for flows.

Service instances are provided to the dependency injection code via factories. The `InjectableFactory` interface must be implemented for any service that should be injected into flows.

An `InjectableFactory` marks a `target` class as the class that the factory returns. This `target` is used when a flow requests an instance of a service, which the factory then provides.

`InjectableFactory` instances are likely to return one of either:

- The platform service instance. This is used for services that are singletons and do not require sandbox or suspending functionality (however not many services will fit this scenario).

- A proxy instance that interacts with platform service. This is used for services that require sandbox or suspending functionality. In this scenario, the proxy will implement the service's public interface while the underlying service will use an internal interface as it may differ from the public one.

`InjectableFactory` instances are discovered via OSGI injection and passed into `DependencyInjectionBuilderFactory` who then passes the instances into the created `DependencyInjectionBuilder`. The same is done for `SingletonSerializeAsToken` instances to register them with the builder and its checkpoint serializer.

The `DependencyInjectionBuilder` returns a `FlowDependencyInjector` that exists within a flow's related sandbox.

The `FlowSandboxService` wraps the logic of getting or creating a sandbox to run a flow in. During this initialisation it creates a `FlowDependencyInjector`, a `CheckpointSerializer` and registers any `discovered platform SingletonSerializeAsToken`s with the serializer.