CORE-2968  Sandbox group context service.
Given a virtual node context of:

* Holding identity
* List of cpks identifiers
* A "sandbox group" 'type' (e.g. worker, flow, etc.)

Return a unique sandbox group, and call any initializer that was passed in
(if not already called).

Lifetime of the sandbox group is managed by the lifetime sandbox group
context service.

A CPK Service interface and a basic bootstrapping implementation
has been introduced such that we can retrieve
a CPK from some implementation given a CPK.Identifier.

Make sure your integration test (OSGI!!) packages are
different to your component (OSGI!!) packages or one
replaces the other, and you cannot start the tests.

InstallService's isRunning needed to be a getter property
not just returning an immutable value as it changes as the
backing coordinator starts and stops.