CORE-3138 Process events outside of flow fibers (#563)
Move the processing of flow events to outside of flow fibers.

This allows better performance as we do not need to deserialize the
flow's checkpoint into a fiber to process incoming events.

The `FlowEventProcessor` provides an implementation of `onNext` that 
takes `FlowEvent`s and uses a `FlowEventPipeline` instance to go through
each processing step.

Currently the primary steps are:

- Pre-processing of the incoming flow event.

- Determining whether a fiber should be run (start, resume or error) or 
continue processing events without interacting with a fiber.

- Post-processing the output `FlowIORequest` the fiber returned when 
suspending, finishing or erroring.

- Post-processing of the flow event received at the start of the 
pipeline.

- Creating a `StateAndEventProcessor.Response` containing the new 
checkpoint and events to send to the messaging bus.

To perform event related processing, `FlowEventHandler` was added. 
Implementations of this interface will handle a single type of event and
can mutate the flow's checkpoint and events to send.

To perform processing based on the last suspending operation (also 
includes finishing and erroring), `FlowRequestHandler` was added. This 
works in the same way as the `FlowEventHandler` but bases its code on 
the flow's output `FlowIORequest`.

General refactors include:
- Renaming `:statemachine` to `:flow-fiber`.
- Renaming `FlowStateMachine` to `FlowFiber`.
- Changing topic names in some setup code (to match the API).
- Removing unused code from classes, such as `HouseKeepingState`, 
`NonSerializableState` and `FlowFiberImpl`.