CORE-1979: Create `Permission` entity using HTTP RPC interface (#611)
This is first part of work on CORE-1979.
It has a bunch of useful changes that permit storing Permission entity to the DB and reply to HTTP RPC call that creation was successful and providing the ID.

However, another PR will be necessary to retrieve Permission by ID, as isolated permissions (i.e. those that are not associated with Roles yet) are not included into in-memory cache on HTTP RPC Gateway side.