CORE-3022: Patterns refactor (#601)
* CORE-3022: Refactor patterns code to separate it from message bus implementations

Split the patterns library into a patterns module and a bus module, divided by the bus API. The patterns module is now common across all bus implementations, while the bus modules implement a smaller API to provide consumers and producers. This change moves the Kafka patterns implementation (as the most complete one) into the patterns implementation, and provides a Kafka bus implementation. The in-memory and DB buses will follow.