CORE-3201 flow mapper tests (#572)
Add integration tests for the flow mapper component.

These rely on OSGi injection to get hold of the required services. Doing this requires the mapper component to set `immediate=true` as otherwise a new instance of the mapper component (and only the mapper component) is injected into every test. The tests have been constructed to work with a singleton service.