CORE-3346 - Populate instance ID properly for simulator receiver/sink mode (#640)
The subscriptions used by the simulator in receiver/sink mode the number of clients as an instance ID, which was leading to conflicts. For instance, in a k8s deployment this led to one instance being fenced from the other (`ProducerFencedException`), which was resulting in very high latencies for the initial messages. Eventually that got resolved and the latencies seemed to go down to normal levels. Most likely the reason is some of the subscriptions eventually gave up and stopped due to crossing the threshold of acceptable errors. For, instance in the simulator logs one could see the following (with producer fenced exception in the stacktrace below):
```
simulator-receiver--725473759: 12:58:48.578 [durable processing thread p2p.in-app-simulator-receiver-p2p.in] ERROR p2p.in-app-simulator-receiver.app-simulator-receiver-1 - Failed to read and process records from topic p2p.in, group p2p.in-app-simulator-receiver, producerClientId app-simulator-receiver-1. Attempts: 1. Closing subscription.
simulator-receiver--725473759: net.corda.messaging.api.exception.CordaMessageAPIFatalException: Fatal error occurred aborting transaction for CordaKafkaProducer with clientId app-simulator-receiver-1
```

Some notes:
* I also removed the producer ID, since it's not needed for subscriptions
* For instance IDs, I chose to use the hashcode of the (instance ID, client ID) string concatenation for simplicity. Given the small number of simulators that will be commonly run, collisions are not likely (and the system would still function properly in the unlikely scenario).