NOTICK: Fix failing on Windows `CordaPackagePersistenceTests` (#642)
When executed on CI with K8s on Windows, the file operations seems to be quite slow/asynchronous. In particular when it comes to deleting of directories and their content, it may lead to having a directory structure in the inconsistent state. Therefore, being a bit more lenient here by retrying deletion of the directories with `eventually` and catching any exception and logging them rather than failing the test.