CORE-2849 - Component that takes RPC config mgmt requests, stores them in DB and republishes to Kafka (#613)
CORE-2849

Defines a new component, components/configuration/configuration-write-service. This component is responsible for taking RPC requests to update the cluster configuration, writing them to the DB, and republishing them onto a Kafka topic to be consumed by the ConfigReader.

This component is plugged into the DBProcessor, which is the principal component of the DB worker.

In tandem, renames the existing components:configuration-write-service-impl/libs:configuration-write/libs:configuration-write-impl to xxx:configuration-publish-xxx, which better reflects their behaviour.