CORE-3010: Remove AMQP Whitelists
A whitelist is a method of bypassing the @CordaSerializable annotation requirements. In Corda 5 we need to be stricter about what we deserialize, to support the sandbox and hosting a shared environment. We need to prevent the serialization system being used to build classes that the sandbox has blocked access to.

Currently there are three ways to allow serialization of a type:

- Add @CordaSerializable
- Define a whitelist that allows the type
- Add a custom serializer for the type

Our platform types are already annotated with @CordaSerializable if we have decided they should be supported by the serialization system. Allowing a whitelist to override the lack of annotation would open up deserialization of any platform type. To close off this, we have decided to remove the whitelist feature.

This PR includes:

- Remove whitelist from serialization-amqp
- Add @CordaSerializable to our test classes that previously relied on an "allow all" whitelist
- Remove whitelist tests
- Update expected exception in some tests to NotSerializableException where the test would previously use a whitelist to get further into serialization code and adding @CordaSerializable would not be the same test
- Add internal custom serializers for DayOfWeek and Month from Java
- Add internal custom serializers for Pair and Unit from Kotlin
- Rename WhitelistBasedTypeModelConfiguration to LocalTypeModelConfigurationImpl
- Replace Mockito with a mock class in PrivateKeyTest so we can target the mock class with a custom serializer
- Regenerate ThrowableEvolutionTests resources because the originals were generated without registering ThrowableSerializer