Core-3183 Gateway fails and closes the connection if partition mapping for session id contains an empty list (porting changes) (#673)
Brief motivation for the change: If partition mapping for a sessionID contained an empty list this produced a failure. This PR was introduced so if this occurs we can fail gracefully. This more graceful fail will give better UX in the interim of preventing the list from being empty/introducing timeouts etc. In this PR the InboundMessageHandler now checks for empty partition list and a warning is logged. The PR also adds changes so for an empty partition we don't initiate sessions and don't reply to session initiations from peers.

Port of #561