NOTICK Add method to sandbox service hasCpks
Added a method to check if the sandbox group service contains the given
set of cpks.  This is so that the flow worker can retry sandbox creation
if the sandbox group service is still receiving cpks from Kafka due to
latency issues.

This implementation simply returns true, as we haven't introduced the
cpk (kakfa) service yet.