CORE-2970 Read/write CPK service (#706)
This revision introduces the initial API of two services that
read and write CPKs to local backing storage.  This has not
been integrated into the code base yet.

In this revision there is an implementation to read and write
to the same folder on disk, rather than Kafka in the next
version.

Added a `WaitingComponent` for the integration test that allows
us to easily wait on the other service components setting their
*status* to UP (not just that `isRunning == true`)

Some projects have also been renamed for consistency and
discoverability (read|write-service|-impl), which we are using
elsewhere in the repository.