CORE-3066 Basic generic chunking library
For some given stream and identifier, we chunk the
stream into Avro `Chunk` messages.

Caller is expected to map the indentifier on the read and
write sides to something useful in an additional message.

Can handle out-of-order chunks.

Final chunk sent is 'zero sized' and contains the final
expected size of the binary and the checksum of the binary.

The 'reader' persists the chunks into a file and returns the
`Path` to that file.

Unit tests cover:

* zero sized messages
* more than one file being received (written)
* out of order chunks
* chunks are trimmed to their actual size
* more in the unit tests themselves

We produce a checksum the stream during the 'read', and the
final file that is 'read' and assembled from the chunks, and
compare the two to ensure integrity.

In this revision there is no handling of incomplete
reassembly of chunks.