CORE-3066 Chunk binaries and pass filename
Addressed PR comments.

Now passing "original" filename as part of the
chunk message.  Updated usage of new Avro object that aligns with
db fields here:

https://github.com/corda/platform-eng-design/blob/master/core/corda-5/corda-5.1/cpi/upload.md#db-entities

The `requestId` is now internal usage only to track the chunks.

Removed random string usage and replaced with UUID for unit tests and
also as the `requestId` which is also the (hopefully) unique name of
the temporary file that the binary is reconstructed into.

Checksum of the incoming binary (stream) and reassembled version are calculated
internally within this library in this revision.