CORE-3066 Basic generic chunking library (#759)
* CORE-3066 Basic generic chunking library

For some given stream and identifier, we chunk the
stream into Avro `Chunk` messages.

Caller is expected to map the indentifier on the read and
write sides to something useful in an additional message.

Can handle out-of-order chunks.

Final chunk sent is 'zero sized' and contains the final
expected size of the binary and the checksum of the binary.

The 'reader' persists the chunks into a file and returns the
`Path` to that file.

Unit tests cover:

* zero sized messages
* more than one file being received (written)
* out of order chunks
* chunks are trimmed to their actual size
* more in the unit tests themselves

We produce a checksum the stream during the 'read', and the
final file that is 'read' and assembled from the chunks, and
compare the two to ensure integrity.

In this revision there is no handling of incomplete
reassembly of chunks.

* CORE-3066 Chunk binaries and pass filename

Now passing "original" filename as part of the
chunk message.  Updated usage of new Avro object that aligns with
db fields here:

https://github.com/corda/platform-eng-design/blob/master/core/corda-5/corda-5.1/cpi/upload.md#db-entities

The `requestId` is now internal usage only to track the chunks.

Removed random string usage and replaced with UUID for unit tests and
also as the `requestId` which is also the (hopefully) unique name of
the temporary file that the binary is reconstructed into.

Checksum of the incoming binary (stream) and reassembled version are calculated
internally within this library in this revision.