CORE-3065  Binary chunking component
Wired into the Db Processor and writes to the database
but does nothing else currently.

Receives `Chunk` objects via RPC, writes them to the database,
and returns a `ChunkAck`.

Database scripts are in `corda-api` in a corresponding PR.
Add missing OSGi export to make the test app correctly
resolve all the chunking parts.

As per PR and discussions with Dries, move the chunking
tables into the config schema for the time being.

Persist a chunk in a transaction and check the parts
vs. expected within the same transaction.  This is to
avoid double sending of a 'final' message if two or
more processors are inserting chunks for the same
`requestId` simultaneously.