CORE-3065  Chunk upload fixes for e2e test
Can now finally upload a given file via the cli:

    curl --insecure -u admin:admin -v -F upload=@some/path/my-1.0.0.jar https://localhost:8888/api/v1/cpi/

We have had to "disable" waiting on the ConfigKeys.MESSAGING_CONFIG as in the
docker-compose set up, we're not receiving that at all, and the whole "cluster"
sits there waiting for something we haven't "injeted" into Kafka.

In other words, just rely on the BOOT_CONFIG for the messaging configuration
for the time being.