CORE-3800 - Add a permission summary topic and integrate into the permission validator and storage reader.
The permission summary topic contains a summary of all permissions for each user, calculated from the data storage in the db worker.

A reconciliation API is added to the PermissionStorageReader to reconcile the permission summaries in the cache with the summaries in the database. Logic has been added to calculate a diff between what is present in the cache and what is in the database. A JPA query performs joins using indexed columns and retrieves a summary of all permissions from roles directly assigned to the User.