CORE-2662: Use different truststore per message (#753)
This change allows the link manager to manage the gateway trusted certificates. The Link Manager will read the certificates from the network map and will publish any new/changed certificate to a compacted topic, which will be consumed by the Gateway and will replace the trust store in the SSL configuration.