NOTICK -  Make Session Manager unit test deterministic (#896)
This change replaces the dependency on time (by using a mock clock and scheduled executor) for the session manager test, so that it becomes completely deterministic and independent of the environment it executes on. We were previously seeing some very rare flakiness only on Windows (that appeared to be due to scheduling/slowness), these changes should remove this flakiness.