CORE-3800 - Integrate permission summary (#891)
The permission summary topic contains a summary of all permissions for each user, calculated using a JPQL query in the DB worker.

A reconciliation API is added to the `PermissionStorageReader` to reconcile the permission summaries in the cache with the summaries in the database. Logic has been added to calculate a diff between what is present in the cache and what is in the database. A JPA query performs joins using indexed columns and retrieves a summary of all permissions from roles directly assigned to the User. Permissions are sorted so order can be relied upon.