NOTICK Change Gateway to always route same initiatorHello to same LinkManager
Before this change we would occassionally see the MAC of the InitiatorHandshake
to fail authentication (when testing a HA deployment). This was caused by the
initiatorHello message being replayed and routed to a different LinkManager.
Meaning we end up with two different responderHello messages. This
causes the mac chack to fail in the InitiatorHandshake.